move: Footer 삭제 및 Header -> Layout 하위로 이동
