--- src/common/components/Footer/style.css.ts
@@ -1,3 +1,0 @@
-import { style, styleVariants } from '@vanilla-extract/css';
-
-import { theme } from 'styles/theme.css';
