+++ src/common/components/Layout/index.tsx
@@ -1,15 +1,13 @@
 import { PropsWithChildren } from 'react';
 import { Outlet } from 'react-router-dom';
 
-import Footer from '@components/Footer';
-import Header from '@components/Header';
+import Header from './Header';
 
 const Layout = ({ children }: PropsWithChildren) => {
   return (
     <>
       <Header />
       <main>{children || <Outlet />}</main>
-      <Footer />
     </>
   );
 };
